Stop Updating the modificationIndex on routes if they haven't changed.
The contract with consumers of routing-api indicates that if the
modification index is newer than the previous, and guids are the same,
the old route should be replaced with the new one.

However, routing-api is updating that index regardless of whether any
changes occurred to the route info (just a heartbeat). This affected
both TCP + HTTP routes.

We no longer increment the modification tag unless the new route is
substantively different from the previous. This also ensures that TCP
and HTTP modification tag processing + event emitting behaviors are
in-line with eachother (this behavior diverged in
7ff498aff2f4a6547ad1e247bccf0ac0e5807e63, when we stopped emitting
upsert events for no-op changes).