Revert "Stop emitting update events when an upsert is empty"
This reverts commit 7ff498aff2f4a6547ad1e247bccf0ac0e5807e63.

We still need to emit an UpdateEvent because the handler has gone
through the action of processing said update event. The catch is that
the Upsert even will *only* be applied _if_ the modification tag of the
route in the routing table is "succeeded by" the new event waiting to be
applied; in the case when the new and existing route mapping match,
nothing will be applied and haproxy will not reload unnecessarily.