Add fields to TCP Route Mappings support TLS backends. (#40)
The HostTLSPort and InstanceId properties have been added to support TLS
backends on TCP routes.

When HostTLSPort is nil, TCP routes can be assumed to be
created by old clients that are unaware of TLS support. When 0, TLS is
explicitly disabled. Positive values indicate the port that the backend
is reachable at via TLS.

InstanceId is the identifier used to verify the backend TLS cert's
hostname against, similar to how Gorouter handles Route Integrity.

Neither field are required for backward compatibility.

Additionally cleanup was done to simplify the various New*() functions
in the TCP Route model, most of which were only ever called from inside
tests.

Signed-off-by: Andy Nguyen <andy-a.nguyen@broadcom.com>