Fixes issue with v6 migration and postgres
Originally in v0.301.0 the v6 migration would fail to migrate on
postgres databases due to existing records without an `instance_id`
column, and adding the column with a not-null constraint. There was
no default value to give the columns so it failed.

On mysql, the migration succeeded, and since mysql is the default
database used in cf-deployment when we test with uptimer, we never
noticed this issue.

Additionally the unit tests for the migration didn't have existing data,
just existing tables.