Fix v6 migration (#59)
* Adds support for TCP Routes with TLS backend ports

Used when a TLS proxy is provided in between tcp-router + the app logic,
to ensure on-wire traffic is always encrypted between the hops.

This reverts commit a4b4dcf60798f0a81b95872aecc0bbdb91e6412f.
This reverts commit 55b4c2d780cd9461255ef1a5b77cf5e706effeaa.
This reverts commit 626a05b402d6a67aa2ed95e8a883d27ee2e91545.
This reverts commit 0bd4442c4d20f88635744306981ad4853c160030.
This reverts commit d961bde4681dab8e4a67ec738f97e97a0b467caa.
This reverts commit 41dd0f0244be665a6f48d5e3c7eee12b56232cf4.
This reverts commit 1628f04444e50b5b35b1fb9308b424e232865738.

* Fixes issue with v6 migration and postgres

Originally in v0.301.0 the v6 migration would fail to migrate on
postgres databases due to existing records without an `instance_id`
column, and adding the column with a not-null constraint. There was
no default value to give the columns so it failed.

On mysql, the migration succeeded, and since mysql is the default
database used in cf-deployment when we test with uptimer, we never
noticed this issue.

Additionally the unit tests for the migration didn't have existing data,
just existing tables.