Update routing-api's db logic to search  for existing records using all unique fields of a tcproutemappingentity
Previously it searched all but sni_hostname + host_tls_port, which
means there were bugs with apps on the same hostIP, both with 0 as the
host_port, but differing host_tls_ports.

Theoretically this is also a bug for SNI routes, which would result in
problems if there was a case where a given backend had a routes to the
same external port where one of them had an SNI hostname defined, and
the other did not. This seems very unlikely (maybe impossible?) to have
occurred in the real world though.

Added tests to ensure that every unique field in the
tcproutemappingentity is accounted for when searching for existing
records.